chatbot/requirements.txt
# requirements.txt

streamlit
langchain
langchain-community
sentence-transformers
faiss-cpu
python-dotenv
langchain-google-genai 
google-generativeai  